"Implement JWT authorization filter"
This commit introduces the JwtAuthorizationFilter, which is responsible for the authorization process of the application.

It verifies the JWT token from the Authorization header of the incoming request. If the token is valid, it will set the authentication in the SecurityContextHolder and the request can proceed. If the token is invalid or not present, it will throw an AuthenticationServiceException.