Implement spring security config (#260)
* "Implement JWT authorization filter"

This commit introduces the JwtAuthorizationFilter, which is responsible for the authorization process of the application.

It verifies the JWT token from the Authorization header of the incoming request. If the token is valid, it will set the authentication in the SecurityContextHolder and the request can proceed. If the token is invalid or not present, it will throw an AuthenticationServiceException.

* Implement JWT Authorization Filter with Spring Security

In this commit, a JWT Authorization Filter is added to the application's Spring Security setup through the WebSecurityConfig class. This filter intercepts incoming requests, checks for a JWT token in the Authorization header, verifies it, and sets the security context if the token is valid.

The secret key for JWT is obtained from the application's properties using the @Value annotation.

* Implemented UserDetails service and custom UserDetails class for Spring Security

* Added tests for Spring Security configuration and updated user statuses in CSV file

* Added tests for Spring Security configuration

* More fields in the User entity and dependencies for pom.xml

* Code Formatting

* Removed UserDetails and Service

And more code formatting

* JwtFilter tests

Changed library for filter

* Improved tests

Finally works

* Code formatting

* Code formatting

Removed some imports, and implemented @Value properly

* Code formatting

Removed not used error handling

* Even more code formatting

* Tests Formatting

Removed some imports, and added tests.

* Added annotations, and improve code formatting

* Update src/main/java/pl/byczazagroda/trackexpensesappbackend/config/WebSecurityConfig.java

* Removed unnecessary @Value annotation

* Improved JwtFilter, and added DeniedHandler

And authenticationEntryPoint

---------

Co-authored-by: Iwona <51120038+Iwona007@users.noreply.github.com>
Co-authored-by: Kacper Poneta <94321164+poneciak57@users.noreply.github.com>