Implement Refresh Token functionality and tests (#279)
* Implement Refresh Token functionality and tests

This commit includes the implementation of refresh tokens functionality in the auth controller, as well as associated tests. Here are the detailed notes on the changes made:

Implementation of the POST /api/auth/refresh endpoint in the auth controller. This function retrieves the refresh token from the request, validates it, and creates a new access token and refresh token. The new refresh token is then added to the response as a cookie.

Implementation of a series of unit tests for the aforementioned function. These tests cover different scenarios such as the case when both the access token and the refresh token are valid, absence of the refresh token, incorrect access token, incorrect refresh token, and the case when the access token has expired.

Moving the RefreshTokenTest tests to the 'controller' package, since these tests are focused on the behavior of the auth controller.

Updating the names of the tests to be more descriptive and adding @DisplayName annotations to each test to make it easier to understand what each test does.

This commit aims to provide refresh tokens functionality, which is essential for the security and convenience of users.

* Code formatting

Code Formatting

* Implement token refreshing 2.0

- removed logic from controller
- access token is now validated
- refreshTokenTest is now refreshTokenIT
- refreshTokenIT is in the integration package
- refreshToken method is better now

* Code formatting

- code is preetier now

* Code formatting

- new error code
- code formatting